allow mock device to use presets and connect via projectkey
